# Conflicts:
#	SIMSclient/Login.txt
#	SIMSclient/src/bussinesslogic/commoditybl/CommodityBL.java
#	SIMSclient/src/po/commodity/CommodityPO.java
#	SIMSserver/src/po/commodity/CommodityPO.java

+++ SIMSclient/src/presentation/financialstaffui/controller/AccountManageController.java
@@ -1,40 +1,35 @@
 package presentation.financialstaffui.controller;
 
-import java.net.URL;
-
 import java.util.ArrayList;
-import java.util.ResourceBundle;
+import java.util.Optional;
 
-import bussiness_stub.AccountBLService_Stub;
 import bussinesslogic.accountbl.AccountController;
 import bussinesslogicservice.accountblservice.AccountBLService;
-import dataenum.Remind;
 import dataenum.ResultMessage;
 import dataenum.findtype.FindAccountType;
 import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableColumn.CellEditEvent;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
-import javafx.scene.control.TableColumn.CellEditEvent;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.util.Callback;
 import presentation.common.EditingCell;
-import presentation.remindui.RemindExistUI;
-import presentation.remindui.RemindPrintUI;
 import vo.accountvo.AccountVO;
 import vo.uservo.UserVO;
 
-public class AccountManageController extends FinancialStaffController implements Initializable{
+public class AccountManageController extends FinancialStaffController{
 
         AccountBLService service = new AccountController();
-        public static final Remind remind = Remind.ACCOUNT;
         ObservableList<AccountVO> list = FXCollections.observableArrayList();
         @FXML
         protected TextField idField;
@@ -62,39 +57,42 @@ public class AccountManageController extends FinancialStaffController implements
 
         @FXML
         public void insert(){
+        	if(idField.getText()==null||nameField.getText()==null||moneyField.getText()==null){
+        		Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好所有的信息");
+        		warning.showAndWait();
+        	}
+        	else{
              AccountVO vo = new AccountVO(idField.getText(), nameField.getText(), Double.parseDouble(moneyField.getText()));
                 ResultMessage message = service.add(vo);
                 Platform.runLater(new Runnable() {
                     public void run() {
                         try {
                         switch(message){
-                        case ILLEGALINPUTNAME:new RemindPrintUI().start(message);break;
-                        case ILLEAGLINPUTDATA:new RemindPrintUI().start(message);break;
-                        case EXISTED:new RemindExistUI().start(remind,true);break;
+                        case EXISTED: Alert existed = new Alert(Alert.AlertType.WARNING,"该账户已存在");
+                                      existed.showAndWait();break;
                         case SUCCESS:list.add(vo);table.setItems(list);initInsert();break;
-                        default:break;
+                        default: Alert error = new Alert(Alert.AlertType.ERROR,message.value);
+                        error.showAndWait();break;
                         }
                         } catch (Exception e) {
                             e.printStackTrace();
                         }
                     }
                 });
+            }
         }
 
 
         @FXML
         public void find(){
+        	if(findingField.getText()==null||findChoice.getValue()==null){
+    			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好查询信息");
+    			warning.showAndWait();
+    		}else{
             ArrayList<AccountVO> accountList = service.find(findingField.getText(),FindAccountType.getFindType(findChoice.getValue()));
                if(accountList==null){
-                   Platform.runLater(new Runnable() {
-                        public void run() {
-                            try {
-                                new RemindPrintUI().start(ResultMessage.ILLEAGLINPUTDATA);
-                            } catch (Exception e) {
-                                e.printStackTrace();
-                            }
-                        }
-                    });
+            	   Alert error = new Alert(Alert.AlertType.WARNING,ResultMessage.NOTFOUND.value);
+                   error.showAndWait();
                }
                else{
                    list.clear();
@@ -102,6 +100,7 @@ public class AccountManageController extends FinancialStaffController implements
                    table.setItems(list);
                    initFind();
                }
+    		}
         }
 
         public void initFind(){
@@ -115,18 +114,16 @@ public class AccountManageController extends FinancialStaffController implements
             moneyField.setText(null);
         }
 
-        @Override
-        public void initialize(URL location, ResourceBundle resources) {
-        	list.clear();
+        public void initData(UserVO user) throws Exception {
+            this.user = user;
+            list.clear();
             list.addAll(service.getAccountList());
             table.setItems(list);
             initFind();
             initInsert();
             edit();
             manageInit();
-
-        }
-
+     }
         public void edit(){
             Callback<TableColumn<AccountVO, String>,
                 TableCell<AccountVO, String>> cellFactory
@@ -159,10 +156,9 @@ public class AccountManageController extends FinancialStaffController implements
                     public void run() {
                         try {
                         switch(message){
-                        case ILLEGALINPUTNAME:new RemindPrintUI().start(message);break;
-                        case ILLEAGLINPUTDATA:new RemindPrintUI().start(message);break;
                         case SUCCESS:break;
-                        default:break;
+                        default: Alert error = new Alert(Alert.AlertType.ERROR,message.value);
+                                       error.showAndWait();break;
                         }
                         } catch (Exception e) {
                             e.printStackTrace();
@@ -198,9 +194,18 @@ public class AccountManageController extends FinancialStaffController implements
                             this.setGraphic(delBtn);
                             delBtn.setOnMouseClicked((me) -> {
                                 AccountVO clickedAccount = this.getTableView().getItems().get(this.getIndex());
-                                service.delete(clickedAccount);
-                                list.remove(clickedAccount);
-                                table.setItems(list);
+                                try {
+	                            	Alert alert = new Alert(AlertType.CONFIRMATION);
+	                            	alert.setContentText("确认删除？");
+	                            	Optional<ButtonType> result = alert.showAndWait();
+	                            	if (result.get() == ButtonType.OK){
+	                            		service.delete(clickedAccount);
+	                            		  list.remove(clickedAccount);
+	      	                              table.setItems(list);
+	                            	}
+								} catch (Exception e) {
+									e.printStackTrace();
+								}
                             });
                         }
                     }
@@ -209,10 +214,5 @@ public class AccountManageController extends FinancialStaffController implements
                 return cell;
             });
         }
-
-        public void initData(UserVO user) throws Exception {
-               this.user = user;
-        }
-
 
 }
