+++ SIMSclient/src/presentation/financialstaffui/controller/ReceiveCheckBillController.java
@@ -1,35 +1,32 @@
 package presentation.financialstaffui.controller;
 
-import java.net.URL;
-
-
 import java.util.ArrayList;
-import java.util.ResourceBundle;
-import bussiness_stub.ReceiptBillBLService_Stub;
+import java.util.Optional;
+
 import bussinesslogic.accountbillbl.ReceiptBillController;
 import bussinesslogicservice.accountbillblservice.ReceiptBillBLService;
 import dataenum.BillState;
 import dataenum.ResultMessage;
 import dataenum.findtype.FindAccountBillType;
-import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
-import presentation.remindui.RemindPrintUI;
 import vo.billvo.financialbillvo.AccountListVO;
 import vo.billvo.financialbillvo.ReceiptBillVO;
 import vo.uservo.UserVO;
 
 
-public class ReceiveCheckBillController extends FinancialStaffController implements Initializable{
+public class ReceiveCheckBillController extends FinancialStaffController{
 
 	ReceiptBillBLService service = new ReceiptBillController();
 	ObservableList<ReceiptBillVO> list = FXCollections.observableArrayList();
@@ -82,43 +79,33 @@ public class ReceiveCheckBillController extends FinancialStaffController impleme
 
 	@FXML
 	public void find(){
-
+		if(findingField.getText()==null||findChoice.getValue()==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好查询信息");
+			warning.showAndWait();
+		}else{
 		ArrayList<ReceiptBillVO> list = service.find(findingField.getText(),FindAccountBillType.getType(findChoice.getValue()));
 	       if(list==null){
-	    	   Platform.runLater(new Runnable() {
-		    	    public void run() {
-		    	        try {
-		    	        	new RemindPrintUI().start(ResultMessage.ILLEAGLINPUTDATA);
-						} catch (Exception e) {
-							e.printStackTrace();
-						}
-		    	    }
-		    	});
+	    	   Alert error = new Alert(Alert.AlertType.WARNING,ResultMessage.NOTFOUND.value);
+               error.showAndWait();
 	       }
 	       else{
 	    	   table.getItems().clear();
 	    	   table.getItems().addAll(list);
 	       }
-
+		}
 	}
 
 
 
 	public void initData(UserVO user) {
 		this.user = user;
-	}
-
-	@Override
-	public void initialize(URL location, ResourceBundle resources) {
 		list.addAll(service.show());
 		table.setItems(list);
 		manageInit();
 		listInit();
 		findChoice.setItems(FXCollections.observableArrayList(FindAccountBillType.BILLID.value,FindAccountBillType.CUSTOMER.value,
 				FindAccountBillType.OPERATOR.value));
-
 	}
-
 
 	public void manageInit(){
 		tableID.setCellValueFactory(
@@ -189,10 +176,10 @@ public class ReceiveCheckBillController extends FinancialStaffController impleme
                         	ResultMessage message = service.commit(clickedItem);
                             if(message == ResultMessage.SUCCESS){
                            	 this.getTableView().getItems().get(this.getIndex()).setState(BillState.COMMITED);
-                                print(ResultMessage.COMMITED);
+                                printInfo(ResultMessage.COMMITED);
                             }
                             else
-                         	   print(message);
+                         	   printWrong(message);
                         });
                     }
                   }
@@ -255,9 +242,18 @@ public class ReceiveCheckBillController extends FinancialStaffController impleme
                         this.setGraphic(delBtn);
                         delBtn.setOnMouseClicked((me) -> {
                         	ReceiptBillVO clickedItem = this.getTableView().getItems().get(this.getIndex());
-                        	service.delete(clickedItem);
-                            list.remove(clickedItem);
-                            table.setItems(list);
+                        	 try {
+	                            	Alert alert = new Alert(AlertType.CONFIRMATION);
+	                            	alert.setContentText("确认删除？");
+	                            	Optional<ButtonType> result = alert.showAndWait();
+	                            	if (result.get() == ButtonType.OK){
+	                            		service.delete(clickedItem);
+	                            		  list.remove(clickedItem);
+	      	                              table.setItems(list);
+	                            	}
+								} catch (Exception e) {
+									e.printStackTrace();
+								}
 
                         });
                     }
