+++ SIMSclient/src/presentation/salestockstaffui/controller/PurchaseCheckBillController.java
@@ -1,10 +1,8 @@
 package presentation.salestockstaffui.controller;
 
-import java.net.URL;
 import java.util.ArrayList;
-import java.util.ResourceBundle;
+import java.util.Optional;
 
-import bussiness_stub.PurchaseBLService_Stub;
 import bussinesslogic.purchasebl.PurchaseController;
 import bussinesslogicservice.purchaseblservice.PurchaseBLService;
 import dataenum.BillState;
@@ -12,25 +10,24 @@ import dataenum.BillType;
 import dataenum.ResultMessage;
 import dataenum.findtype.FindPurchaseType;
 import dataenum.findtype.FindSalesType;
-import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.PropertyValueFactory;
-import presentation.remindui.RemindPrintUI;
 import vo.billvo.purchasebillvo.PurchaseVO;
-import vo.billvo.salesbillvo.SalesVO;
 import vo.commodityvo.CommodityItemVO;
 import vo.uservo.UserVO;
 
-public class PurchaseCheckBillController extends SaleStockStaffController implements Initializable {
+public class PurchaseCheckBillController extends SaleStockStaffController {
 
 	PurchaseBLService service = new PurchaseController();
 	ObservableList<PurchaseVO> list = FXCollections.observableArrayList();
@@ -126,33 +123,27 @@ public class PurchaseCheckBillController extends SaleStockStaffController implem
 
 	@FXML
 	public void find(){
-
+		if(findingField.getText()==null||findChoice.getValue()==null){
+			Alert warning = new Alert(Alert.AlertType.WARNING,"请填写好查询信息");
+			warning.showAndWait();
+		}
+		else{
 		ArrayList<PurchaseVO> list = service.find(findingField.getText(),FindSalesType.getType(findChoice.getValue()));
 	       if(list==null){
-	    	   Platform.runLater(new Runnable() {
-		    	    public void run() {
-		    	        try {
-		    	        	new RemindPrintUI().start(ResultMessage.ILLEAGLINPUTDATA);
-						} catch (Exception e) {
-							e.printStackTrace();
-						}
-		    	    }
-		    	});
+	    	   Alert error = new Alert(Alert.AlertType.WARNING,ResultMessage.NOTFOUND.value);
+               error.showAndWait();
 	       }
 	       else{
 	    	   table.getItems().clear();
 	    	   table.getItems().addAll(list);
 	       }
+	    }
 
 	}
 
 	public void initData(UserVO user,BillType type) {
 		this.user = user;
 		this.type = type;
-	}
-
-	@Override
-	public void initialize(URL location, ResourceBundle resources) {
 		list.addAll(service.show());
 		table.setItems(list);
 		manageInit();
@@ -235,10 +226,10 @@ public class PurchaseCheckBillController extends SaleStockStaffController implem
                         	ResultMessage message = service.submit(clickedItem);
                             if(message == ResultMessage.SUCCESS){
                            	 this.getTableView().getItems().get(this.getIndex()).setState(BillState.COMMITED);
-                                print(ResultMessage.COMMITED);
+                                printInfo(ResultMessage.COMMITED);
                             }
                             else
-                         	   print(message);
+                         	   printWrong(message);
                         });
                     }
                   }
@@ -301,9 +292,14 @@ public class PurchaseCheckBillController extends SaleStockStaffController implem
                         this.setGraphic(delBtn);
                         delBtn.setOnMouseClicked((me) -> {
                         	PurchaseVO clickedItem = this.getTableView().getItems().get(this.getIndex());
-                        	service.delete(clickedItem);
-                            list.remove(clickedItem);
-                            table.setItems(list);
+                        	 Alert alert = new Alert(AlertType.CONFIRMATION);
+							alert.setContentText("确认删除？");
+							Optional<ButtonType> result = alert.showAndWait();
+							if (result.get() == ButtonType.OK){
+								service.delete(clickedItem);
+								  list.remove(clickedItem);
+							      table.setItems(list);
+							}
 
                         });
                     }
